Move pipeline_components inside pipelines
