--- pipeline_components/aiplatform/aiplatform/conftest.py
@@ -1,1 +1,0 @@
-# Required for Python imports during pytest unit tests
