+++ pipelines/pipelines/kfp_components/aiplatform/aiplatform/import_model_evaluation.py
@@ -13,7 +13,6 @@
 # limitations under the License.
 
 from kfp.v2.dsl import Input, Model, Metrics, component, Dataset
-from pathlib import Path
 from typing import NamedTuple
 
 
@@ -22,7 +21,6 @@ from typing import NamedTuple
     packages_to_install=[
         "google-cloud-aiplatform>=1.24.1",
     ],
-    output_component_file=str(Path(__file__).with_suffix(".yaml")),
 )
 def import_model_evaluation(
     model: Input[Model],
