+++ pipelines/pipelines/kfp_components/aiplatform/aiplatform/update_best_model.py
@@ -14,13 +14,11 @@ from typing import NamedTuple
 # limitations under the License.
 
 from kfp.v2.dsl import Input, Model, component
-from pathlib import Path
 
 
 @component(
     base_image="python:3.7",
     packages_to_install=["google-cloud-aiplatform==1.24.1"],
-    output_component_file=str(Path(__file__).with_suffix(".yaml")),
 )
 def update_best_model(
     challenger: Input[Model],
