+++ pipelines/pipelines/kfp_components/aiplatform/tests/test_update_best_model.py
@@ -20,7 +20,7 @@ def test_model_batch_predict(tmpdir):
     Asserts model_batch_predict successfully creates requests given different arguments.
     """
 
-    from aiplatform.update_best_model.component import update_best_model
+    from aiplatform.update_best_model import update_best_model
 
     mock_model = Model(uri=tmpdir, metadata={"resourceName": ""})
     mock_message = {"metrics": {"rmse": 0.01}}
