+++ pipelines/pipelines/kfp_components/bigquery/.python-version
