+++ pipelines/pipelines/kfp_components/bigquery/bigquery/extract_bq_to_dataset.py
@@ -13,13 +13,11 @@
 # limitations under the License.
 
 from kfp.v2.dsl import Dataset, Output, component
-from pathlib import Path
 
 
 @component(
     base_image="python:3.7",
     packages_to_install=["google-cloud-bigquery==2.30.0"],
-    output_component_file=str(Path(__file__).with_suffix(".yaml")),
 )
 def extract_bq_to_dataset(
     bq_client_project_id: str,
