+++ pipelines/pipelines/tensorflow/training/pipeline.py
@@ -18,7 +18,7 @@ import pathlib
 
 from kfp.v2 import compiler, dsl
 from pipelines import generate_query
-from pipelines.components import (
+from pipelines.kfp_components import (
     extract_bq_to_dataset,
     bq_query_to_table,
     update_best_model,
